% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.filter.R
\name{fun.filter}
\alias{fun.filter}
\title{Auxiliar function to remove duplicated or zero variance regressors}
\usage{
fun.filter(X)
}
\description{
Auxiliar function to remove duplicated regressors.
}
\keyword{internal}
