% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.poly.R
\name{fun.poly}
\alias{fun.poly}
\title{Performs polynomial to use in GMDH Combinatorial.}
\usage{
fun.poly(X, G = 2)
}
\arguments{
\item{X}{matrix containing independent variables in the model.}

\item{G}{polynomial. Default is 2, original Ivakhnenko polynomial.
0: normal regression.
1: only interaction terms.}
}
\value{
A matrix with n columns corresponding with Ivakhnenko polynomial
}
\description{
Performs polynomial to use in GMDH Combinatorial.
}
\keyword{internal}
