% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.svd.R
\name{fun.svd}
\alias{fun.svd}
\title{Perform svd regression}
\usage{
fun.svd(x, y)
}
\arguments{
\item{y}{vector or matrix containing dependent variable in the model.}

\item{X}{matrix containing independent variables in the model.}
}
\description{
Calculates svd regression.
}
\keyword{internal}
