% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.svd_2.R
\name{fun.svd_2}
\alias{fun.svd_2}
\title{Perform svd regression}
\usage{
fun.svd_2(x, y, x.test, y.test)
}
\arguments{
\item{y}{vector or matrix containing dependent variable in the model.}

\item{X}{matrix containing independent variables in the model.}
}
\description{
Calculates svd regression.
}
\keyword{internal}
