% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.svd_3.R
\name{fun.svd_3}
\alias{fun.svd_3}
\title{Perform svd regression}
\usage{
fun.svd_3(x, y)
}
\arguments{
\item{x}{matrix containing independent variables in the model.}

\item{y}{vector or matrix containing dependent variable in the model.}
}
\description{
Calculates svd regression.
}
\keyword{internal}
