% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi.twice_1.R
\name{gmdh.combi.twice_1}
\alias{gmdh.combi.twice_1}
\title{GMDH COMBI-TWICE auxiliar functions}
\usage{
gmdh.combi.twice_1(X, y, G)
}
\description{
Performs auxiliar tasks to predict.mia
}
\keyword{internal}
