% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi.twice_2.R
\name{gmdh.combi.twice_2}
\alias{gmdh.combi.twice_2}
\title{GMDH COMBI-TWICE auxiliar function}
\usage{
gmdh.combi.twice_2(X, y, G, x.test, y.test)
}
\description{
Build a regression model performing GMDH Twice-Multilayered Combinatorial (TMC) with external criteria.
}
\keyword{internal}
