% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_1.R
\name{gmdh.combi_1}
\alias{gmdh.combi_1}
\title{GMDH COMBI auxiliar functions}
\usage{
gmdh.combi_1(X, y, G = 2)
}
\description{
Build a regression model performing GMDH Combinatorial with PRESS criteria.
}
\keyword{internal}
