% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_2.R
\name{gmdh.combi_2}
\alias{gmdh.combi_2}
\title{GMDH COMBI auxiliar functions}
\usage{
gmdh.combi_2(X, y, G = 2, x.test, y.test)
}
\description{
Build a regression model performing GMDH Combinatorial with external criteria.
}
\keyword{internal}
