% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_3.R
\name{gmdh.combi_3}
\alias{gmdh.combi_3}
\title{GMDH COMBI auxiliar functions}
\usage{
gmdh.combi_3(X, y, G = 2)
}
\description{
Build a regression model performing GMDH Combinatorial with ICOMP criteria.
}
\keyword{internal}
