% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.svd.R
\name{predict.svd}
\alias{predict.svd}
\title{Predict svd regression}
\usage{
\method{predict}{svd}(object, newdata, ...)
}
\arguments{
\item{object}{An object model from fun.svd an object class 'svd'}

\item{newdata}{matrix containing dependent variable in the model, wich the predictions are calculated.}
}
\description{
Calculates svd regression model predictions for new data.
}
\keyword{internal}
