% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi.twice_1.R
\name{gmdh.combi.twice_1}
\alias{gmdh.combi.twice_1}
\title{GMDH COMBI-TWICE auxiliar function}
\usage{
gmdh.combi.twice_1(X, y, G)
}
\description{
Build a regression model performing GMDH Twice-Multilayered Combinatorial (TMC) with PRESS criteria.
}
\keyword{internal}
