% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.N.R, R/predict.neurona.R
\name{predict.neurona}
\alias{predict.neurona}
\title{Predict neurona object}
\usage{
\method{predict}{neurona}(object, newdata, ...)

\method{predict}{neurona}(object, newdata, ...)
}
\arguments{
\item{object}{An object model from mia an object class 'neurona'}

\item{newdata}{matrix containing dependent variable in the model, wich the predictions are calculated.}
}
\description{
Calculates neurona model predictions for new data.

Performs auxiliar tasks to predict.mia
}
\examples{

NULL

}
\keyword{internal}
